requirements.txt
pillow
opencv-python
numpy
flask
werkzeug
